Added a env variable to control data deletion on tx table

+++ src/worker/listeners/deleteProcessedTx.ts
@@ -1,4 +1,5 @@
 import cron from "node-cron";
+import { env } from "../../utils/env";
 import { deleteTx } from "../tasks/deleteTx";
 
 const CLEAR_QUEUED_TX_CRON_SCHEDULE = "0 0 */2 * * *";
@@ -6,6 +7,8 @@ const CLEAR_QUEUED_TX_CRON_SCHEDULE = "0 0 */2 * * *";
 // Deletes successfully processed transactions which were queued 24 hrs ago.
 export const deleteProcessedTx = async () => {
   cron.schedule(CLEAR_QUEUED_TX_CRON_SCHEDULE, async () => {
-    await deleteTx();
+    if (env.PRUNE_TRANSACTIONS) {
+      await deleteTx();
+    }
   });
 };
